LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/scroll-timeline-axis-writing-mode-expected.txt


FAIL Initial axis assert_equals: expected "175px" but got "0px"
FAIL Vertical axis assert_equals: expected "175px" but got "0px"
FAIL Horizontal axis assert_equals: expected "125px" but got "0px"
FAIL Block axis in horizontal writing-mode assert_equals: expected "175px" but got "0px"
FAIL Inline axis in horizontal writing-mode assert_equals: expected "125px" but got "0px"
FAIL Block axis in vertical writing-mode assert_equals: expected "125px" but got "0px"
FAIL Inline axis in vertical writing-mode assert_equals: expected "175px" but got "0px"
